+++ projects/savia-web/src/__tests__/test-utils.ts
@@ -1,0 +1,4 @@
+import i18n from '../../src/locales'
+
+/** Global mount options that include i18n plugin for all component tests */
+export const globalPlugins = { global: { plugins: [i18n] } }